+++ CHANGELOG.md
@@ -24,3 +24,9 @@
 # 0.0.6
 
 * Add ability to set the output medium.
+
+# 0.0.7
+
+* Start informing the server of the client version.
+* Use simplifed `transcript` messages.
+* Expose `sendData` for bleeding edge use cases.